Seeds and migration updates

+++ database/migrations/2018_05_13_232640_create_bookings_table.php
@@ -19,7 +19,7 @@ class CreateBookingsTable extends Migration
             $table->integer('room_id');
             $table->date('check_in');
             $table->date('check_out');
-            $table->float('price');
+            $table->double('price', 8, 2);
             $table->timestamps();
         });
     }
